pyproject.toml
[project]
name = "archipelagomanualchapterchallenge"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0"
]
